refactor: add a normalizer plugin
