+++ packages/rich-text/src/plugins/Break/index.ts
@@ -1,2 +1,4 @@
+export type { SoftBreakRule } from '@udecode/plate-break';
+
 export * from './createSoftBreakPlugin';
 export * from './createExitBreakPlugin';
