+++ packages/rich-text/src/plugins/EmbeddedEntityBlock/index.tsx
@@ -4,12 +4,11 @@ import { BLOCKS } from '@contentful/rich-text-types';
 import { KeyboardHandler, HotkeyPlugin } from '@udecode/plate-core';
 import { Transforms } from 'slate';
 import { getNodeEntryFromSelection } from '../../helpers/editor';
-import { CustomElement } from '../../types';
+import { RichTextPlugin, CustomElement } from '../../types';
 import { LinkedEntityBlock } from './LinkedEntityBlock';
 import { selectEntityAndInsert } from './Util';
 import { FieldExtensionSDK } from '@contentful/app-sdk';
 import noop from 'lodash/noop';
-import { RichTextPlugin } from '../types';
 
 export { EmbeddedEntityBlockToolbarIcon as ToolbarIcon } from './ToolbarIcon';
