+++ packages/rich-text/src/plugins/Marks/Bold.tsx
@@ -9,7 +9,7 @@ import { FormatBoldIcon } from '@contentful/f36-icons';
 
 import { ToolbarButton } from '../shared/ToolbarButton';
 import { useContentfulEditor } from '../../ContentfulEditorProvider';
-import { RichTextPlugin } from '../types';
+import { RichTextPlugin } from '../../types';
 
 interface ToolbarBoldButtonProps {
   isDisabled?: boolean;
