+++ packages/rich-text/src/plugins/Marks/Underline.tsx
@@ -7,7 +7,7 @@ import { MARKS } from '@contentful/rich-text-types';
 import { FormatUnderlinedIcon } from '@contentful/f36-icons';
 import { ToolbarButton } from '../shared/ToolbarButton';
 import { useContentfulEditor } from '../../ContentfulEditorProvider';
-import { RichTextPlugin } from '../types';
+import { RichTextPlugin } from '../../types';
 
 interface ToolbarUnderlineButtonProps {
   isDisabled?: boolean;
