+++ packages/rich-text/src/plugins/Marks/index.ts
@@ -2,7 +2,7 @@ import { createBoldPlugin } from './Bold';
 import { createCodePlugin } from './Code';
 import { createItalicPlugin } from './Italic';
 import { createUnderlinePlugin } from './Underline';
-import { RichTextPlugin } from '../types';
+import { RichTextPlugin } from '../../types';
 
 export const createMarksPlugin = (): RichTextPlugin => ({
   key: 'Marks',
