+++ packages/rich-text/src/plugins/Normalizer/createNormalizerPlugin.ts
@@ -1,0 +1,7 @@
+import { RichTextPlugin } from '../../types';
+import { withNormalizer } from './withNormalizer';
+
+export const createNormalizerPlugin = (): RichTextPlugin => ({
+  key: 'NormalizerPlugin',
+  withOverrides: withNormalizer,
+});
