+++ packages/rich-text/src/plugins/Normalizer/index.ts
@@ -1,0 +1,2 @@
+export * from './types';
+export * from './createNormalizerPlugin';
