+++ packages/rich-text/src/plugins/Paste/index.tsx
@@ -1,7 +1,7 @@
 import { PlateEditor } from '@udecode/plate-core';
 import * as sanitizers from './sanitizers';
 import flow from 'lodash/flow';
-import { RichTextPlugin } from '../types';
+import { RichTextPlugin } from '../../types';
 
 const MIME_TYPE_HTML = 'text/html';
