+++ packages/rich-text/src/plugins/Table/components/Cell.tsx
@@ -1,0 +1,33 @@
+import { css } from 'emotion';
+import * as React from 'react';
+import * as Slate from 'slate-react';
+import tokens from '@contentful/f36-tokens';
+import { TableCell } from '@contentful/rich-text-types';
+
+import { TableActions } from './TableActions';
+
+const style = css`
+  border: 1px solid ${tokens.gray400};
+  padding: 10px 12px;
+  min-width: 48px;
+  position: relative;
+
+  div:last-child {
+    margin-bottom: 0;
+  }
+`;
+
+export const Cell = (props: Slate.RenderElementProps) => {
+  const isSelected = Slate.useSelected();
+
+  return (
+    <td
+      {...props.attributes}
+      // may include `colspan` and/or `rowspan`
+      {...(props.element.data as TableCell['data'])}
+      className={style}>
+      {isSelected && <TableActions />}
+      {props.children}
+    </td>
+  );
+};
