+++ packages/rich-text/src/plugins/Table/helpers.ts
@@ -10,37 +10,13 @@ import {
 import { getAbove, getChildren, isFirstChild, isAncestorEmpty } from '@udecode/plate-core';
 import { BLOCKS } from '@contentful/rich-text-types';
 
-import {
-  isBlockSelected,
-  getNodeEntryFromSelection,
-  getAncestorPathFromSelection,
-} from '../../helpers/editor';
+import { isBlockSelected, getAncestorPathFromSelection } from '../../helpers/editor';
 import { CustomElement } from '../../types';
-
-/**
- * Sets the UI focus to the first cell of the selected table.
- *
- * This is necessary because udecode's plugin sets the focus to the
- * _last_ cell of the table after selection, which is weird.
- */
-function moveToFirstCellFromSelectedTable(editor) {
-  const [, tablePath] = getNodeEntryFromSelection(editor, BLOCKS.TABLE);
-  if (!tablePath) return;
-  const firstTableCellPath = tablePath.concat([
-    0, // -> row
-    0, // -> cell
-    0, // -> paragraph
-    0, // -> text node
-  ]);
-  const anchor = { path: firstTableCellPath, offset: 0 };
-  Transforms.setSelection(editor, { anchor, focus: anchor });
-}
 
 export function insertTableAndFocusFirstCell(editor: PlateEditor): void {
   // FIXME: a table should only be allowed at root level. Currently this
   // code adds it at any level
   insertTable(editor, { header: true });
-  moveToFirstCellFromSelectedTable(editor);
   replaceEmptyParagraphWithTable(editor);
 }
